requirements.txt
playwright
pandas
requests
pillow
openai
urllib3<2
python-dotenv
pyyaml
